move 3d tiles get records to catalog api
